(feat) implement tracker level filtering in #4779
