+++ cmd/nuclei/issue-tracker-config.yaml
@@ -1,3 +1,8 @@
+# global allow/deny list. this will affect both exporters
+# as well as issue trackers. you can filter trackers with
+# a tracker level filter on top of an exporter by setting
+# allow-list/deny-list per tracker.
+#
 #allow-list:
 #  severity: high, critical
 #deny-list:
@@ -17,6 +22,15 @@
 #  project-name: test-project
 #  # issue-label is the label of the created issue type
 #  issue-label: bug
+#  # allow-list sets a tracker level filter to only create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  allow-list:
+#    severity: high, critical
+#    tags: network
+#  # deny-list sets a tracker level filter to never create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  deny-list:
+#    severity: low
 #  # duplicate-issue-check flag to enable duplicate tracking issue check.
 #  duplicate-issue-check: false
 #
@@ -32,6 +46,15 @@
 #  project-name: "1234"
 #  # issue-label is the label of the created issue type
 #  issue-label: bug
+#  # allow-list sets a tracker level filter to only create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  allow-list:
+#    severity: high, critical
+#    tags: network
+#  # deny-list sets a tracker level filter to never create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  deny-list:
+#    severity: low
 #
 # Gitea contains configuration options for a gitea issue tracker
 #gitea:
@@ -47,6 +70,15 @@
 #  issue-label: bug
 #  # severity-as-label (optional) adds the severity as a label of the created issue
 #  severity-as-label: true
+#  # allow-list sets a tracker level filter to only create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  allow-list:
+#    severity: high, critical
+#    tags: network
+#  # deny-list sets a tracker level filter to never create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  deny-list:
+#    severity: low
 #  # duplicate-issue-check (optional) flag to enable duplicate tracking issue check
 #  duplicate-issue-check: false
 #
@@ -71,6 +103,15 @@
 #  # SeverityAsLabel (optional) sends the severity as the label of the created issue
 #  # User custom fields for Jira Cloud instead
 #  severity-as-label: true
+#  # allow-list sets a tracker level filter to only create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  allow-list:
+#    severity: high, critical
+#    tags: network
+#  # deny-list sets a tracker level filter to never create issues for templates with
+#  # these severity labels or tags (does not affect exporters. set those globally)
+#  deny-list:
+#    severity: low
 #  # Whatever your final status is that you want to use as a closed ticket - Closed, Done, Remediated, etc
 #  # When checking for duplicates, the JQL query will filter out status's that match this.
 #  # If it finds a match _and_ the ticket does have this status, a new one will be created.
